requirements.txt
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
idna==3.11
invoke==2.2.1
Jinja2==3.1.6
junos-eznc==2.7.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
napalm==5.1.0
ncclient==0.7.0
netaddr==1.3.0
netmiko==4.6.0
netutils==1.15.1
ntc_templates==9.0.0
numpy==2.4.3
pandas==3.0.1
paramiko==4.0.0
pycparser==3.0
pyeapi==1.0.4
Pygments==2.19.2
PyNaCl==1.6.2
pyparsing==3.3.2
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
ruamel.yaml==0.19.1
scp==0.15.0
setuptools==82.0.1
six==1.17.0
tabulate==0.10.0
textfsm==2.1.0
transitions==0.9.3
ttp==0.10.1
ttp_templates==0.5.1
typing_extensions==4.15.0
urllib3==2.6.3
yamlloader==1.6.0
